requirements.txt
catkin==0.8.10
cycler==0.11.0
fonttools==4.37.1
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
kiwisolver==1.4.4
matplotlib==3.5.3
message-filters==1.15.14
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
picamera==1.13
Pillow==9.2.0
pkg-resources==0.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
RPi.GPIO==0.7.1
six==1.16.0
StereoVision==1.0.4
topic-tools==1.15.14
typing-extensions==4.3.0
